% Auto-generated: do not edit by hand
\name{''NestedDataCRUD}

\alias{''NestedDataCRUD}

\title{NestedDataCRUD component}

\description{

}

\usage{
''NestedDataCRUD(id=NULL, dataSource=NULL, keyField=NULL, keyHeader=NULL, valueField=NULL,
valueHeader=NULL, childrenField=NULL, nodeLabel=NULL, leafLabel=NULL,
height=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataSource}{List of named lists. }

\item{keyField}{Character. }

\item{keyHeader}{Character. }

\item{valueField}{Character. }

\item{valueHeader}{Character. }

\item{childrenField}{Character. }

\item{nodeLabel}{Character. }

\item{leafLabel}{Character. }

\item{height}{Numeric. }
}
